requirements.txt
asgiref==3.8.1
build==1.2.1
click==8.1.7
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
packaging==24.1
psycopg2==2.9.9
PyJWT==2.8.0
pyproject_hooks==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
setuptools==70.0.0
sqlparse==0.5.0
wheel==0.43.0
